Attempt at Angular Acceleration
Trying to redesign the ship and the functions that govern how it moves.
Made progress but causes a stack overflow because of too many nested
picture transforms like translate and rotate. Not sure how since every
time we call updatePicture there is just two transforms. Confirmed,
however, that it only takes 15 updatePictures to cause the overflow
which is close to the depth of 32 stack frames.